(fix) Parser: Norminette OK, create expand_utils.c
