+++ src/parsing/expand.c
@@ -6,77 +6,62 @@
 /*   By: gpollast <gpollast@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/15 09:11:35 by gpollast          #+#    #+#             */
-/*   Updated: 2025/09/18 19:03:46 by gpollast         ###   ########.fr       */
+/*   Updated: 2025/09/18 19:51:05 by gpollast         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-size_t	get_env_var_len(char *word)
+static t_list	*add_content_to_list(t_list **lst, char *content)
 {
-	size_t	i;
-
-	if (!word)
-		return (0);
-	i = 0;
-	while (word[i] && word[i] != ' ' && word[i] != '$' && word[i] != '\'')
-	{
-		i++;
-	}
-	return (i);
+	if (!content)
+		return (ft_lstclear(lst, free), NULL);
+	ft_lstadd_back(lst, ft_lstnew(content));
+	return (*lst);
 }
 
-char	*my_getenv(t_msh *msh, char *word)
+static t_list	*process_var(t_list **lst, char **ptr)
 {
-	int	i;
-	int	len;
+	char	*content;
+	char	*var_ptr;
+	size_t	var_len;
 
-	i = 0;
-	len = get_env_var_len(word + 1);
-	while (msh->env[i])
-	{
-		if (!ft_strncmp(msh->env[i], word + 1, len) && msh->env[i][len] == '=')
-		{
-			return (msh->env[i] + len + 1);
-		}
-		i++;
-	}
-	return (NULL);
+	var_ptr = ft_strchr(*ptr, '$');
+	content = ft_substr(*ptr, 0, var_ptr - *ptr);
+	if (!add_content_to_list(lst, content))
+		return (NULL);
+	var_len = get_env_var_len(var_ptr + 1);
+	content = ft_substr(var_ptr, 0, var_len + 1);
+	if (!add_content_to_list(lst, content))
+		return (NULL);
+	*ptr = var_ptr + var_len + 1;
+	return (*lst);
 }
 
 t_list	*split_env_var(char *s)
 {
-    t_list	*lst;
-    char	*content;
-    char	*var_ptr;
-    char	*ptr;
+	t_list	*lst;
+	char	*content;
+	char	*ptr;
 
-    lst = NULL;
-    ptr = s;
-    while (*ptr)
-    {
-        var_ptr = ft_strchr(ptr, '$');
-        if (var_ptr)
-        {
-            content = ft_substr(ptr, 0, var_ptr - ptr);
-            if (!content)
-                return (ft_lstclear(&lst, free), NULL);
-            ft_lstadd_back(&lst, ft_lstnew(content));
-            content = ft_substr(var_ptr, 0, get_env_var_len(var_ptr + 1) + 1);
-            if (!content)
-                return (ft_lstclear(&lst, free), NULL);
-            ptr = var_ptr + get_env_var_len(var_ptr + 1) + 1;
-        }
-        else
-        {
-            content = ft_substr(ptr, 0, ft_strlen(ptr));
-            if (!content)
-                return (ft_lstclear(&lst, free), NULL);
-            ptr = ptr + ft_strlen(ptr);
-        }
-        ft_lstadd_back(&lst, ft_lstnew(content));
-    }
-    return (lst);
+	lst = NULL;
+	ptr = s;
+	while (*ptr)
+	{
+		if (ft_strchr(ptr, '$'))
+		{
+			if (!process_var(&lst, &ptr))
+				return (NULL);
+		}
+		else
+		{
+			content = ft_substr(ptr, 0, ft_strlen(ptr));
+			if (!add_content_to_list(&lst, content))
+				return (NULL);
+			ptr = ptr + ft_strlen(ptr);
+		}
+	}
+	return (lst);
 }
 
 char	*replace_env_var(t_msh *msh, char *s)
@@ -106,29 +91,29 @@ char	*replace_env_var(t_msh *msh, char *s)
 
 char	*expand(t_msh *msh, char *s)
 {
-    char	*res;
-    char	*tmp;
-    t_list	*lst;
-    t_list	*current;
+	char	*res;
+	char	*tmp;
+	t_list	*lst;
+	t_list	*current;
 
-    if (!s)
-        return (NULL);
-    res = malloc(sizeof(char));
-    if (!res)
-        return (NULL);
-    res[0] = 0;
-    lst = split_env_var(s);
-    current = lst;
-    while (current)
-    {
-        tmp = res;
-        current->content = replace_env_var(msh, ((char *)current->content));
-        res = ft_strjoin(res, current->content);
-        if (!res)
-            return (ft_lstclear(&lst, free), free(tmp), NULL);
-        free(tmp);
-        current = current->next;
-    }
-    ft_lstclear(&lst, free);
-    return (res);
+	if (!s)
+		return (NULL);
+	res = malloc(sizeof(char));
+	if (!res)
+		return (NULL);
+	res[0] = 0;
+	lst = split_env_var(s);
+	current = lst;
+	while (current)
+	{
+		tmp = res;
+		current->content = replace_env_var(msh, ((char *)current->content));
+		res = ft_strjoin(res, current->content);
+		if (!res)
+			return (ft_lstclear(&lst, free), free(tmp), NULL);
+		free(tmp);
+		current = current->next;
+	}
+	ft_lstclear(&lst, free);
+	return (res);
 }
